LayoutTests/fast/css/apple-system-control-colors-expected.txt

-apple-system-header-text : rgb(0, 0, 0)
-apple-system-text-background : rgb(255, 255, 255)
-apple-system-alternate-selected-text : rgb(255, 255, 255)
-apple-system-label : rgb(0, 0, 0)
-apple-system-secondary-label : rgb(0, 0, 0)
-apple-system-tertiary-label : rgb(0, 0, 0)
-apple-system-quaternary-label : rgb(0, 0, 0)
-apple-system-grid : rgb(204, 204, 204)
current-color with inherited -apple-system-label : rgb(0, 0, 0)
